updated dtos to make creation of summoners possible added cascade options to rank, summoner and division

+++ src/summoners/summoner.controller.ts
@@ -1,39 +1,39 @@
 import {
-  Controller,
-  Get,
-  Post,
   Body,
-  Patch,
-  Param,
+  Controller,
   Delete,
+  Get,
   HttpException,
   HttpStatus,
+  Param,
+  Patch,
+  Post,
+  Request,
+  UnauthorizedException,
   UseGuards,
   ValidationPipe
 } from '@nestjs/common'
-import { SummonerService } from './summoner.service'
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard'
+import { JwtAuthenticatedRequest } from 'src/auth/jwt.strategy'
 import { CreateSummonerDto } from './dto/create-summoner.dto'
 import { UpdateSummonerDto } from './dto/update-summoner.dto'
-import {
-  Summoner,
-  SummonerOmittingPasswordHash
-} from './entities/summoner.entity'
-import { JwtAuthGuard } from 'src/auth/jwt-auth.guard'
+import { SummonerOmittingPasswordHash } from './entities/summoner.entity'
+import { SummonerService } from './summoner.service'
 
 @Controller('summoner')
 export class SummonerController {
   constructor(private readonly summonerService: SummonerService) {}
 
   @Post()
-  create(
+  async create(
     @Body(new ValidationPipe()) createSummonerDto: CreateSummonerDto
-  ): Promise<Summoner | undefined> {
-    return this.summonerService.create(createSummonerDto)
+  ): Promise<SummonerOmittingPasswordHash | undefined> {
+    return await this.summonerService.create(createSummonerDto)
   }
 
   @Get()
-  findAll(): Promise<Summoner[] | undefined> {
-    return this.summonerService.findAll()
+  async findAll(): Promise<SummonerOmittingPasswordHash[] | undefined> {
+    return await this.summonerService.findAll()
   }
 
   @Get(':id')
@@ -51,18 +51,32 @@ export class SummonerController {
 
   @UseGuards(JwtAuthGuard)
   @Patch(':id')
-  update(
+  async update(
     @Param('id') id: number,
-    @Body() updateSummonerDto: UpdateSummonerDto
-  ): Promise<Summoner | undefined> {
-    //TODO - Check if the summoner is the same as the one in the JWT or user is admin. Else throw 403.
-    return this.summonerService.update(id, updateSummonerDto)
+    @Body() updateSummonerDto: UpdateSummonerDto,
+    @Request() { summoner }: JwtAuthenticatedRequest
+  ): Promise<SummonerOmittingPasswordHash | undefined> {
+    const summonerFromDB = await this.summonerService.findOne(id)
+
+    // Autorize admins and summoners
+    if (!summoner.isAdmin && summoner.id !== summonerFromDB?.id)
+      throw new UnauthorizedException()
+
+    return await this.summonerService.update(id, updateSummonerDto)
   }
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: number): Promise<void> {
-    //TODO - Check if the summoner is the same as the one in the JWT or user is admin. Else throw 403.
-    return this.summonerService.remove(+id)
+  async remove(
+    @Param('id') id: number,
+    @Request() { summoner }: JwtAuthenticatedRequest
+  ): Promise<void> {
+    const summonerFromDB = await this.summonerService.findOne(id)
+
+    // Autorize admins and summoners
+    if (!summoner.isAdmin && summoner.id !== summonerFromDB?.id)
+      throw new UnauthorizedException()
+
+    return await this.summonerService.remove(+id)
   }
 }
